Nearly a rewrite of the module
This began with simply supporting a new platform.  As the deployment of
this module increased, it became clear that much of what this module was
intended to do was a) not a problem I had or want to have, and b) much
of the noise that existed in this repository was to support that
complexity.

This work backs out the use cases that I don't intend to support as this
module gets deployed in the relatively modern infrastructure that I am
now working to build out.  As such, this targets only recent versions of
supported operating systems, and intends to maintain feature parity on
all systems with a reduced set of features.

While making these changes, puppet4 is supported exlusively, and even
then, only the most recent versions that exist on the platforms I'm
currently involved with.

Gunicorn is dropped.  This seemed site specific and didn't really belong
in this module.  Gunicorn can still be installed in a virtualenv or with
pip, or both, but vhost management for a specific application server
didn't seem to belong.

There are still documentation updates, and classes that need a good
combing, but the core functionality of this module remains intact.