Update the pip installation on FreeBSD
Drop the hacky exec to install pip for python3 versions on FreebSD.
This work replaces the exec with a package installation matching the
version of the default system python.

Setting DEFAULT_VERSION for python on a FreeBSD system allows the py-pip
package to be built into a name like py35-pip when the default version
is set to 35.  This work ensures that we use the package in place of the
curl.

There may be nuances around trying to instal multiple versions of pip
on the same system.