Fix testing for FreeBSD
Without this change, the FreeBSD tests are not actually enabled due to
the facterdb handling of the FreeBSD version in the metadata.  Here we
remove the point release from the FreeBSD version in the metadata,
enabling the tests, and then fix the tests that are broken.

Here also we replace the pip installation requirement for virtualenv
with usage of py$x-virtualenv packages on FreeBSD.  The packages will
now be used as python continues to improve on FreeBSD systems.