Begin updating data logic
Here we remove the function that provided the name of the virtualenv
package to use and move the logic from that function into hiera data.
Going forward, this will require that each version of python supported
on a given platform have extra data associated with it for the
virtualenv command and package name to use.

This should make it simpler to test existing platforms, and to support
to new platforms.

Here also, we begin testing virtualenv support, which was not done
previously.  This became clear when a recent upgrade to OpenBSD 6.0
revealed an issue with the way the virtualenv was created.

Initial data for Archlinux is also added here, though no tests are here
to prove it works.  More to come here.