Relocate pip data into hiera
This work follows the same pattern of moving the information about the
command and package name for the selected version of python into hiera
and out of the native function.

Initial Archlinux support.

Drop distribute support during virtualenv creation.  Distribute is a
fork of setuptools, but the changes have been merged back.  There should
be no need to support distribute in this module.