Improve default version data for FreeBSD
I've recently stopped setting version information on FreeBSD and realized that
this code doesn't handle the default version of python on FreeBSD very well.
Currently that version is 2.7, so here we just set the version_defaults hash
that is used in absence of a specific version to reference the values from the
'27' version of python.