python::pip: Implement group support
In my long term effort to merge this module with voxpupuli/python, I'm
adding a parameter that's present in the voxpupuli module. Now it's
possible to set a group for installed pips and not only the owner.

this is breaking change for people that currently have owner != root. I
plan to propose more PRs in the future, so please don't release this one
yet.